Commit dejar punto de restauracion

+++ htdocs/App/Controllers/Perfil.php
@@ -2,6 +2,7 @@
 namespace App\Controllers;
 
 defined("APPPATH") OR die("Access denied");
+//isset($_SESSION['usuario']) OR die("Access denied2");
 
 use \Core\View,
 \App\Models\User as Users,
@@ -21,31 +22,37 @@ class Perfil{
 		if(!isset($_SESSION['usuario'])){
 			View::render('login');
 		}else{
-			define('USUARIO',$_SESSION['usuario']->getNombre());
 			View::render('perfil');
 		}
 	}
 	// Usado para una llamada Post de ajax
 	public function modificarPass(){
+		// Se convierten los indices en variables
 		extract($_POST);
-		//en caso de enviar algun campo vacio
-		if(empty($oldPass)||empty($newPass)){
+
+		//en caso de recibir algún campo vacio
+		if(empty($oldPass)||empty($newPass)||empty($cNewPass)){
 			exit("faltan campos por rellenar");
 		}
+
 		//Se codifica la contraseña y se verifica que coincida con la actual
 		$oldPass=md5($oldPass);
-		echo $oldPass==$_SESSION['usuario']->getPass()?"":exit("Contraseña antigua incorrecta");
-		//Validacion
+		$oldPass==$_SESSION['usuario']->getPass()?:exit("Contraseña antigua incorrecta");
+		
+		//Validacion Contraseña por Expresion regular
 		if(!preg_match("/(?=^.{6,}$)((?=.*\d)|(?=.*\W+))(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$/",$newPass)){
 			exit("Requisitos Minimos:<br>1 Mayuscula<br>1 Minuscual<br>1 Digitio<br>6 Caracteres o mas");
 		}
+		//Si la contraseña de confirmación no coincide
 		if($newPass!=$cNewPass){
 			exit("Contraseña de confirmación no valida");
 		}
-
+		//Finalmente se codifica la contraseña
 		$newPass=md5($newPass);
 
-		if(UserAdmin::update("SET pass = '$newPass' WHERE id= $id AND pass = '$oldPass'")==true){
+		//Se actualiza la contraseña en la base de datos y en la session en caso de que se actualice correctamente
+		$ok=UserAdmin::update("SET pass = '$newPass' WHERE id= $id AND pass = '$oldPass'");
+		if($ok==1){
 			echo "Contraseña modificada con exito";
 			$_SESSION['usuario']->setPass($newPass);
 		}
@@ -53,7 +60,9 @@ class Perfil{
 	//Genera la vista de gestion de usuarios
 	public function gestion(){
 		if($_SESSION['usuario']->getAdmin()){
+			//Constante para verificar que solo se pueda acceder a la vista desde aqui
 			define("ADMIN",1);
+			
 			$queryUsers=UserAdmin::getAll();
 			$usuarios=array();
 			//Se almacena cada usuario en la clase Perfil y todos en un array
@@ -63,15 +72,26 @@ class Perfil{
 			//se envian todos los usuarios con su corresponiente clase perfil
 			View::set("usuarios",$usuarios);
 			View::render("gestionUsuarios");
+		}else{
+			//Si un usuario no administrador intenta acceder
+			die("Acceso Denegado");
 		}
-		die("Acceso Denegado");
 	}
 	public function resetPass(){
-		extract($_POST);
-		UserAdmin::update("SET pass = '".md5("Portol1") . "' where id = $id");
+		if($_SESSION['usuario']->getAdmin()){
+			//Constante para verificar que solo se pueda acceder a la vista desde aqui
+			extract($_POST);
+			UserAdmin::update("SET pass = '".md5("Portol1") . "' where id = $id");
+		}
 	}
 	public function delUser(){
-		extract($_POST);
-		UserAdmin::delete($id);
+		if($_SESSION['usuario']->getAdmin()){
+			extract($_POST);
+			if($admin==0){
+				UserAdmin::delete($id);
+			}else{
+				echo "No puedes eliminar un administrador";
+			}
+		}
 	}
 }
